+++ src/mbgl/renderer/debug_bucket.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <mbgl/tile/tile_data.hpp>
 #include <mbgl/map/mode.hpp>
 #include <mbgl/geometry/debug_font_buffer.hpp>
 #include <mbgl/geometry/vao.hpp>
@@ -8,6 +7,7 @@
 
 namespace mbgl {
 
+class OverscaledTileID;
 class PlainShader;
 
 namespace gl {
