+++ src/mbgl/text/get_anchors.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <mbgl/geometry/anchor.hpp>
-#include <mbgl/tile/geometry_tile.hpp>
+#include <mbgl/tile/geometry_tile_data.hpp>
 #include <mbgl/util/math.hpp>
 
 namespace mbgl {
