+++ src/mbgl/text/quads.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <mbgl/text/glyph.hpp>
-#include <mbgl/tile/geometry_tile.hpp>
+#include <mbgl/tile/geometry_tile_data.hpp>
 
 #include <vector>
