+++ src/mbgl/tile/geojson_tile.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <mbgl/tile/geometry_tile_data.hpp>
+#include <mbgl/tile/geometry_tile.hpp>
 
 namespace mapbox {
 namespace geojsonvt {
@@ -14,13 +14,12 @@ namespace style {
 class UpdateParameters;
 }
 
-class GeoJSONTileData : public GeometryTileData {
+class GeoJSONTile : public GeometryTile {
 public:
-    GeoJSONTileData(const OverscaledTileID&,
-                   std::string sourceID,
-                   const style::UpdateParameters&,
-                   mapbox::geojsonvt::GeoJSONVT*);
-    ~GeoJSONTileData();
+    GeoJSONTile(const OverscaledTileID&,
+                std::string sourceID,
+                const style::UpdateParameters&,
+                mapbox::geojsonvt::GeoJSONVT*);
 
     void setNecessity(Necessity) final;
 };
