+++ src/mbgl/tile/tile_cache.cpp
@@ -1,5 +1,5 @@
 #include <mbgl/tile/tile_cache.hpp>
-#include <mbgl/tile/tile_data.hpp>
+#include <mbgl/tile/tile.hpp>
 
 #include <cassert>
 
@@ -17,21 +17,21 @@ void TileCache::setSize(size_t size_) {
     assert(orderedKeys.size() <= size);
 }
 
-void TileCache::add(const OverscaledTileID& key, std::unique_ptr<TileData> data) {
-    if (!data->isRenderable() || !size) {
+void TileCache::add(const OverscaledTileID& key, std::unique_ptr<Tile> tile) {
+    if (!tile->isRenderable() || !size) {
         return;
     }
 
-    // insert new or query existing data
-    if (tiles.emplace(key, std::move(data)).second) {
-        // remove existing data key
+    // insert new or query existing tile
+    if (tiles.emplace(key, std::move(tile)).second) {
+        // remove existing tile key
         orderedKeys.remove(key);
     }
 
-    // (re-)insert data key as newest
+    // (re-)insert tile key as newest
     orderedKeys.push_back(key);
 
-    // purge oldest key/data if necessary
+    // purge oldest key/tile if necessary
     if (orderedKeys.size() > size) {
         get(orderedKeys.front());
     }
@@ -39,19 +39,19 @@ void TileCache::add(const OverscaledTileID& key, std::unique_ptr<TileData> data)
     assert(orderedKeys.size() <= size);
 }
 
-std::unique_ptr<TileData> TileCache::get(const OverscaledTileID& key) {
+std::unique_ptr<Tile> TileCache::get(const OverscaledTileID& key) {
 
-    std::unique_ptr<TileData> data;
+    std::unique_ptr<Tile> tile;
 
     auto it = tiles.find(key);
     if (it != tiles.end()) {
-        data = std::move(it->second);
+        tile = std::move(it->second);
         tiles.erase(it);
         orderedKeys.remove(key);
-        assert(data->isRenderable());
+        assert(tile->isRenderable());
     }
 
-    return data;
+    return tile;
 }
 
 bool TileCache::has(const OverscaledTileID& key) {
