--- src/mbgl/tile/tile_data.cpp
@@ -1,37 +1,0 @@
-#include <mbgl/tile/tile_data.hpp>
-#include <mbgl/tile/tile_data_observer.hpp>
-#include <mbgl/renderer/debug_bucket.hpp>
-#include <mbgl/util/string.hpp>
-
-namespace mbgl {
-
-static TileDataObserver nullObserver;
-
-TileData::TileData(const OverscaledTileID& id_)
-    : id(id_), observer(&nullObserver) {
-}
-
-TileData::~TileData() = default;
-
-void TileData::setObserver(TileDataObserver* observer_) {
-    observer = observer_;
-}
-
-void TileData::setTriedOptional() {
-    triedOptional = true;
-    observer->onNeedsRepaint();
-}
-
-void TileData::dumpDebugLogs() const {
-    Log::Info(Event::General, "TileData::id: %s", util::toString(id).c_str());
-    Log::Info(Event::General, "TileData::renderable: %s", isRenderable() ? "yes" : "no");
-    Log::Info(Event::General, "TileData::complete: %s", isComplete() ? "yes" : "no");
-}
-
-void TileData::queryRenderedFeatures(
-        std::unordered_map<std::string, std::vector<Feature>>&,
-        const GeometryCoordinates&,
-        const TransformState&,
-        const optional<std::vector<std::string>>&) {}
-
-} // namespace mbgl
