+++ src/mbgl/tile/tile_observer.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <exception>
+
+namespace mbgl {
+
+class Tile;
+
+class TileObserver {
+public:
+    virtual ~TileObserver() = default;
+
+    virtual void onTileLoaded(Tile&, bool /*isNewTile*/) {}
+    virtual void onTileError(Tile&, std::exception_ptr) {}
+    virtual void onNeedsRepaint() {}
+};
+
+} // namespace mbgl
